scrapper/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.1.31
charset-normalizer==3.4.1
fake-useragent==2.1.0
frozenlist==1.7.0
greenlet==3.2.1
grpcio==1.71.0
idna==3.10
jalali_core==1.0.0
jdatetime==5.2.0
multidict==6.6.3
playwright==1.52.0
propcache==0.3.2
protobuf==3.12.4
pyee==13.0.0
pyperclip==1.9.0
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
soupsieve==2.7
typing_extensions==4.13.2
urllib3==2.3.0
yarl==1.20.1